+++ hypothesis_geometry/core/utils.py
@@ -5,12 +5,7 @@ from typing import (Callable,
                     Sequence,
                     Tuple)
 
-from ground.base import (Context,
-                         Orientation)
-from ground.hints import Point
-
 from .hints import (Domain,
-                    Orienteer,
                     Range)
 
 flatten = chain.from_iterable
@@ -24,22 +19,13 @@ def pairwise(iterable: Iterable[Domain]) -> Iterable[Tuple[Domain, Domain]]:
         element = next_element
 
 
-def to_contour_orienteer(context: Context) -> Callable[[Sequence[Point]],
-                                                       Iterable[Orientation]]:
-    return partial(_to_contour_orientations, context.angle_orientation)
-
-
-def _to_contour_orientations(orienteer: Orienteer,
-                             vertices: Sequence[Point]
-                             ) -> Iterable[Orientation]:
-    return (orienteer(vertices[index], vertices[index - 1],
-                      vertices[(index + 1) % len(vertices)])
-            for index in range(len(vertices)))
-
-
 def apply(function: Callable[..., Range],
           args: Iterable[Domain]) -> Range:
     return function(*args)
+
+
+def cut(values: Domain, limit: int) -> Domain:
+    return values[:limit] if limit < len(values) else values
 
 
 def pack(function: Callable[..., Range]
@@ -50,3 +36,5 @@ def pack(function: Callable[..., Range]
 def sort_pair(pair: Sequence[Domain]) -> Tuple[Domain, Domain]:
     first, second = pair
     return (first, second) if first < second else (second, first)
+
+
